Added centering script
Added a python script for performing centering on 4DSTEM data suitable for use on clusters etc. The script is based on estimating linear shifts from centre of mass within certain regions of the diffraction patterns - see the help of the script for more details on how to use the script.